Rename CombinedDisabilityRatingCard.jsx
